+++ internal/sparkwingruntime/runner.go
@@ -16,5 +16,5 @@ func WithRunner(ctx context.Context, r *sparkwing.RunnerInfo) context.Context {
 	if ctx == nil {
 		ctx = context.Background()
 	}
-	return context.WithValue(ctx, sparkwing.RuntimePlumbing.Runner, r)
+	return context.WithValue(ctx, sparkwing.RuntimePlumbing.Keys.Runner, r)
 }